requirements.txt
Flask==2.0.1
numpy==1.19.5
Pillow==8.3.1
tensorflow==2.6.0
tensorflow-hub==0.12.0